# api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
  /api/user:
    get:
      tags:
      - user
      operationId: getOwnUser
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User'
    put:
      tags:
      - user
      operationId: updateOwnUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUser'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User'
    delete:
      tags:
      - user
      operationId: deleteUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Auth'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: integer
                format: int64
  /api/user/activate:
    put:
      tags:
      - user
      operationId: activateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivateUser'
        required: true
      responses:
        "200":
          description: OK
  /api/schedule:
    get:
      tags:
      - schedule
      operationId: getAllSchedules
      parameters:
      - name: plantId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        required: true
        schema:
          minimum: 0
          type: integer
          format: int32
      - name: pageSize
        in: query
        required: true
        schema:
          maximum: 100
          minimum: 5
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageSchedule'
    put:
      tags:
      - schedule
      operationId: updateSchedule
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Schedule'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Schedule'
    post:
      tags:
      - schedule
      operationId: createSchedule
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSchedule'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Schedule'
...
  /api/user/register:
    post:
      tags:
      - user
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
        required: true
      responses:
        "200":
          description: OK
...
  /api/schedule/{id}:
    get:
      tags:
      - schedule
      operationId: getOneSchedule
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Schedule'
    delete:
      tags:
      - schedule
      operationId: deleteSchedule
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: integer
                format: int64
...
components:
  schemas:
    UpdateUser:
      required:
      - email
      type: object
      properties:
        email:
          maxLength: 255
          minLength: 5
          type: string
        firstName:
          maxLength: 50
          minLength: 0
          type: string
        lastName:
          maxLength: 50
          minLength: 0
          type: string
        oldPassword:
          maxLength: 255
          minLength: 8
          type: string
        password:
          maxLength: 255
          minLength: 8
          type: string
    User:
      required:
      - email
      - username
      type: object
      properties:
        username:
          maxLength: 25
          minLength: 5
          type: string
        email:
          maxLength: 255
          minLength: 5
          type: string
        firstName:
          maxLength: 50
          minLength: 0
          type: string
        lastName:
          maxLength: 50
          minLength: 0
          type: string
        password:
          maxLength: 255
          minLength: 8
          type: string
    ActivateUser:
      required:
      - token
      - username
      type: object
      properties:
        username:
          maxLength: 25
          minLength: 5
          type: string
        token:
          type: string
    EventType:
      required:
      - id
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        schedulable:
          type: boolean
    PlantBasic:
      required:
      - id
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          maxLength: 100
          minLength: 0
          type: string
    Schedule:
      required:
      - autoUpdate
      - id
      - next
      - plant
      - type
      type: object
      properties:
        id:
          type: integer
          format: int64
        plant:
          $ref: '#/components/schemas/PlantBasic'
        type:
          $ref: '#/components/schemas/EventType'
        repeatEvery:
          maximum: 365
          minimum: 1
          type: integer
          format: int32
        autoUpdate:
          type: boolean
        next:
          type: string
          format: date
...
    CreateUser:
      required:
      - email
      - password
      - username
      type: object
      properties:
        username:
          maxLength: 25
          minLength: 5
          type: string
        email:
          maxLength: 255
          minLength: 5
          type: string
        firstName:
          maxLength: 50
          minLength: 0
          type: string
        lastName:
          maxLength: 50
          minLength: 0
          type: string
        password:
          maxLength: 255
          minLength: 8
          type: string
    CreateSchedule:
      required:
      - autoUpdate
      - next
      - plant
      - type
      type: object
      properties:
        plant:
          $ref: '#/components/schemas/PlantBasic'
        type:
          $ref: '#/components/schemas/EventType'
        repeatEvery:
          maximum: 365
          minimum: 1
          type: integer
          format: int32
        autoUpdate:
          type: boolean
        next:
          type: string
          format: date
...
    Auth:
      required:
      - password
      - username
      type: object
      properties:
        username:
          maxLength: 25
          minLength: 0
          type: string
        password:
          maxLength: 255
          minLength: 0
          type: string
...
    PageSchedule:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        number:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Schedule'
        sort:
          $ref: '#/components/schemas/Sort'
        first:
          type: boolean
        last:
          type: boolean
        pageable:
          $ref: '#/components/schemas/Pageable'
        empty:
          type: boolean
    Pageable:
      type: object
      properties:
        paged:
          type: boolean
        unpaged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/Sort'
    Sort:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
...